# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Rumah Sakit",
        "description": "API ini digunakan untuk mengelola data pasien, dokter, konsultasi, dan lainnya.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server API Lokal"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/konsultasi": {
            "get": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Menampilkan semua data konsultasi",
                "operationId": "3a57cd90c3a84dda05941bf7cf6bfed4",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil semua data konsultasi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Menyimpan data konsultasi baru",
                "operationId": "882765e8d04b2f3530f9875e7ced8f1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pasien_id",
                                    "dokter_id",
                                    "jadwal_id",
                                    "tanggal_konsultasi",
                                    "status"
                                ],
                                "properties": {
                                    "pasien_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dokter_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "jadwal_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "tanggal_konsultasi": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dijadwalkan",
                                            "Selesai",
                                            "Dibatalkan"
                                        ],
                                        "example": "Dijadwalkan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Konsultasi berhasil disimpan"
                    },
                    "500": {
                        "description": "Gagal menyimpan konsultasi"
                    }
                }
            }
        },
        "/api/konsultasi/{id}": {
            "get": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Menampilkan detail konsultasi",
                "operationId": "7a69db09f31e76446f2b632982b57081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID konsultasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail konsultasi"
                    },
                    "404": {
                        "description": "Konsultasi tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Memperbarui data konsultasi",
                "operationId": "114a0d580c72f0e0eee9a44207654b5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID konsultasi yang akan diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pasien_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dokter_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "jadwal_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "tanggal_konsultasi": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dijadwalkan",
                                            "Selesai",
                                            "Dibatalkan"
                                        ],
                                        "example": "Selesai"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konsultasi berhasil diperbarui"
                    },
                    "404": {
                        "description": "Konsultasi tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Menghapus data konsultasi",
                "operationId": "cd47f6a62273e14dad622c39c34c46f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID konsultasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konsultasi berhasil dihapus"
                    },
                    "404": {
                        "description": "Konsultasi tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}